Add files via upload
Our proposed solution is to provide an application that can store and centrally manage this data. The application will have a web component and mobile app. Patients or designated caregivers can enter the patient’s medical history details, which the app will store securely. The data can then be accessed by
any medical provider or caregiver the patient authorizes. Providers will be able to update the medical history during an appointment. Patients will also be able to view information about the provider. The system will incorporate medical IoT devices such as FitBit and other smart devices that can connect to the application and upload their data, which the application will use to generate reports and trending
information for statistical analysis.